feat: Add keynote abstract
This commit adds the Quarto configuration file for the keynote submission. The configuration specifies the type of manuscript, the article file, and the HTML format. It also includes the index.qmd file for the keynote submission, which contains metadata such as categories, title, author information, keywords, and abstract. The abstract discusses the ethics and politics of digitization in the context of digital historical literacy.

Date: 05-10-2024
Date-modified: 05-30-2024

Co-authored-by: Gerben Zaagsma <gerbenzaagsma@users.noreply.github.com>